+++ src/components/index.js
@@ -2,5 +2,6 @@ import Text from './Text/Text';
 import Icon from './Icon/Icon';
 import Header from './Header/Header';
 import Pressable from './Pressable/Pressable';
+import InboxName from './InboxName/InboxName';
 
-export { Text, Icon, Header, Pressable };
+export { Text, Icon, Header, Pressable, InboxName };
