+++ src/screens/Conversation/components/ConversationEmptyList/ConversationEmptyList.stories.js
@@ -1,8 +1,8 @@
 import { storiesOf } from '@storybook/react-native';
 import React from 'react';
-import ConversationItemLoader from './ConversationItemLoader';
+import ConversationEmptyList from './ConversationEmptyList';
 import CenterView from 'components/StoryBookView';
 
-storiesOf('Conversation list loader', module)
+storiesOf('Conversation empty list', module)
   .addDecorator(getStory => <CenterView>{getStory()}</CenterView>)
-  .add('loading view', () => <ConversationItemLoader />);
+  .add('loading view', () => <ConversationEmptyList />);
