+++ src/screens/Conversation/components/ConversationEmptyList/ConversationEmptyList.style.js
@@ -1,0 +1,9 @@
+import { StyleSheet } from 'react-native';
+
+export default theme => {
+  return StyleSheet.create({
+    container: {
+      minHeight: 64,
+    },
+  });
+};
