+++ storybook/storyLoader.js
@@ -4,25 +4,23 @@
 // https://github.com/elderfo/react-native-storybook-loader.git
 
 function loadStories() {
-  require('../src/components/Header/Header.stories');
-  require('../src/components/Icon/Icon.stories');
-  require('../src/components/Pressable/Pressable.stories');
-  require('../src/components/Text/Text.stories');
-  require('../src/screens/Conversation/components/ConversationItemLoader/ConversationItemLoader.stories');
-  require('../src/components/InboxName/InboxName.stories');
   require('../src/components/Empty/Empty.stories');
   require('../src/components/FilterButton/FilterButton.stories');
+  require('../src/components/Header/Header.stories');
+  require('../src/components/Icon/Icon.stories');
+  require('../src/components/InboxName/InboxName.stories');
+  require('../src/components/Pressable/Pressable.stories');
+  require('../src/components/Text/Text.stories');
 }
 
 const stories = [
-  '../src/components/Header/Header.stories',
-  '../src/components/Icon/Icon.stories',
-  '../src/components/Pressable/Pressable.stories',
-  '../src/components/Text/Text.stories',
-  '../src/screens/Conversation/components/ConversationItemLoader/ConversationItemLoader.stories',
-  '../src/components/InboxName/InboxName.stories',
   '../src/components/Empty/Empty.stories',
   '../src/components/FilterButton/FilterButton.stories',
+  '../src/components/Header/Header.stories',
+  '../src/components/Icon/Icon.stories',
+  '../src/components/InboxName/InboxName.stories',
+  '../src/components/Pressable/Pressable.stories',
+  '../src/components/Text/Text.stories',
 ];
 
 module.exports = {
